build: 升级至 Spring Boot 3.4.4 并迁移到 Jakarta EE
- 将 Spring Boot 从 2.7.18 升级到 3.4.4
- 将所有 javax.* 导入迁移到 jakarta.*（如 javax.servlet, javax.annotation）
- 更新相关依赖版本（Redisson, MyBatis-Plus, Caffeine等）以兼容 Spring Boot 3
- 调整配置文件中的 Redis 配置项前缀（spring.redis.* -> spring.data.redis.*）
- 新增 Spring AI 相关依赖以支持 AI 功能
- 修正部分方法调用（如 count() 返回类型转换）

+++ hm-dianping/src/main/java/com/hmdp/service/impl/FollowServiceImpl.java
@@ -10,10 +10,11 @@ import com.hmdp.mapper.FollowMapper;
 import com.hmdp.service.IFollowService;
 import com.hmdp.service.IUserService;
 import com.hmdp.utils.UserHolder;
+import jodd.typeconverter.impl.IntegerConverter;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -37,7 +38,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
     @Override
     public Result queryNotFollow(Long id) {
     Long userId=UserHolder.getUserId();
-    Integer count=query().eq("user_id",userId).eq("follow_user_id",id).count();
+    Integer count= Math.toIntExact(query().eq("user_id", userId).eq("follow_user_id", id).count());
         return Result.ok(count>0);
     }
